burglar sensor
This code is an Arduino program that controls an LED and a buzzer based on the signal from a sensor. The sensor is connected to Pin 9, and when it receives a HIGH signal, both the buzzer (on Pin 11) and the LED (on Pin 8) are activated for 200 milliseconds. If the sensor gives a LOW signal, both the buzzer and the LED turn off, and the system waits for 200 milliseconds.