IR remote project
This code detects the keystrokes of an IR remote control using an infrared (IR) receiver and adjusts the colors of an RGB LED (red, green, blue) based on these keystrokes. A different control key is assigned for each color; when certain keys are pressed, the corresponding color value increases or decreases, and one key resets the LED. The code shows which key was pressed by printing the HEX code of the detected IR signal via the serial port.