Bubble machine
This repository contains the control code for a simple bubble machine that works using an Arduino. It produces bubbles by rotating in soapy water through servo motor. In addition, it helps to produce bubbles more effectively by controlling LEDs and a fan.

Required Components:
Arduino Board (Uno, Mega, Nano etc.)
Servo Motor
LEDs
Transistor or MOSFET (for fan control)
Resistors (for LEDs)
Breadboard and Jumper Cables

Setup:
Connect the servo motor to the Arduino's pin number 9.
Connect the LEDs to pins 3, 4 and 5 respectively.
Make the necessary connections to control the fan via transistor or MOSFET.

The Principle of Operation of the Code:
The code generates bubbles by rotating the servo motor in soapy water. LEDs and fan support the bubble production process. The servo motor rotates at certain angles, stirring the soapy water and causing bubbles to form. The speed of the fan is controlled using the analogWrite function.