Radar device
This code measures the distance of environmental objects using a servo motor and the HC-SR04 ultrasonic distance sensor and sends this distance through the serial port while rotating in 15-degree steps. If an object is detected at a distance of 10 cm or closer, "Danger!" gives the warning; otherwise it sends the message "Stable".